+++ src/main/java/com/example/internhub_be/repository/SkillRepository.java
@@ -3,5 +3,5 @@ package com.example.internhub_be.repository;
 import com.example.internhub_be.domain.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillRepository extends JpaRepository<Skill, Integer> {
+public interface SkillRepository extends JpaRepository<Skill, Long> {
 }
